Convert values to integers

+++ PopPUNK/network.py
@@ -530,8 +530,8 @@ def network_to_edges(prev_G_fn, rlist, adding_qq_dists = False,
             G_df.columns = ['weights','src','dst'] # This appears to be a bug in cugraph v0.19
             edge_weights = G_df['weights'].to_arrow().to_pylist()
         G_df.rename(columns={'src': 'source','dst': 'destination'}, inplace=True)
-        old_source_ids = G_df['source'].to_arrow().to_pylist()
-        old_target_ids = G_df['destination'].to_arrow().to_pylist()
+        old_source_ids = G_df['source'].astype('int32').to_arrow().to_pylist()
+        old_target_ids = G_df['destination'].astype('int32').to_arrow().to_pylist()
     else:
         # get the source and target nodes
         old_source_ids = gt.edge_endpoint_property(prev_G, prev_G.vertex_index, "source")
